+++ packages/cli/src/util/deploy/process-deployment.ts
@@ -20,8 +20,6 @@ import { displayBuildLogs } from '../logs';
 import { progress } from '../output/progress';
 import ua from '../ua';
 import output from '../../output-manager';
-import getProjectByNameOrId from '../projects/get-project-by-id-or-name';
-import { ProjectNotFound } from '../errors-ts';
 
 function printInspectUrl(
   inspectorUrl: string | null | undefined,
@@ -124,8 +122,6 @@ export default async function processDeployment({
     abortController?.abort();
     output.stopSpinner();
   }
-
-  let rollingRelease: ProjectRollingRelease | undefined;
 
   try {
     for await (const event of createDeployment(clientOptions, requestBody)) {