+++ packages/cli/test/unit/commands/deploy/index.test.ts
@@ -687,15 +687,6 @@ describe('deploy', () => {
       });
       deployment = useDeployment({ creator: user, state: 'BUILDING' });
       deployment.aliasAssigned = false;
-
-      client.scenario.get(`/v9/projects/:id`, (_req, res) => {
-        res.json({
-          ...defaultProject,
-          name: 'node',
-          id: 'QmbKpqpiUqbcke',
-        });
-      });
-
       client.scenario.post(`/v13/deployments`, (req, res) => {
         res.json(
           res.json({
@@ -1252,13 +1243,6 @@ describe('deploy', () => {
         const user = useUser();
         client.scenario.get(`/v9/projects/:id`, (_req, res) => {
           return res.status(404).json({});
-        });
-
-        client.scenario.get(`/v9/projects`, (_req, res) => {
-          return res.json({
-            projects: [],
-            pagination: {},
-          });
         });
 
         client.scenario.post(`/v1/projects`, (req, res) => {