+++ packages/cli/test/unit/commands/redeploy/index.test.ts
@@ -306,10 +306,6 @@ function initRedeployTest({
       return;
     }
   );
-  client.scenario.get(`/v9/projects/:projectNameOrId`, (_req, res) => {
-    res.json(project);
-    return;
-  });
 
   return {
     project,